refactor: useInputs의 에러 상태를 관리하는 useValidations 커스텀 훅 구현
